Loyalty page integration

+++ lib/models/transaction.dart
@@ -1,7 +1,7 @@
 class Transaction {
   String? id;
   String name;
-  double amount;
+  int amount;
   String category;
   String date;
 
@@ -16,7 +16,7 @@ class Transaction {
   Transaction.fromJson(Map<String, dynamic> json)
     : id = json['_id'] as String?,
       name = json['name'] as String,
-      amount = json['amount'] as double,
+      amount = json['amount'] as int,
       category = json['category'] as String,
       date = json['date'] as String;
 }
